+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Some refactoring for the ``interpreter.Interpreter`` class:
+
+- (Removed) ``_select_eventless_transition`` which is a special case of ``_select_transition``.
+- (Added) ``_select_event``, extracted from ``execute_once``.
+- (Added) ``_filter_transitions``, extracted from ``_select_transition``.
+- (Changed) ``_execute_step`` is now ``_apply_step``.
+- (Changed) ``_compute_stabilization_step`` is now ``_create_stabilization_step`` and accepts a list of state names
+- (Changed) ``_compute_transitions_step`` is now ``_create_steps``.
+
 0.20.5 (2016-04-14)
 -------------------